+++ affinidi-messaging-helpers/src/mediator_administration/account_management/acl_management.rs
@@ -172,8 +172,12 @@ async fn _modify_acl_flags(
         ),
         ("access_list self-change?", acls.get_self_manage_list()),
         (
-            "queue limits self-change?",
-            acls.get_self_manage_queue_limit(),
+            "queue send limits self-change?",
+            acls.get_self_manage_send_queue_limit(),
+        ),
+        (
+            "queue receive limits self-change?",
+            acls.get_self_manage_receive_queue_limit(),
         ),
     ];
 
@@ -186,7 +190,7 @@ async fn _modify_acl_flags(
         .unwrap();
 
     // convert the selection to an array of bools
-    let mut flags = [false; 18];
+    let mut flags = [false; 19];
     for s in selection {
         flags[s] = true;
     }
@@ -211,7 +215,8 @@ async fn _modify_acl_flags(
     let _ = new_acls.set_create_invites(flags[12], flags[13], true);
     let _ = new_acls.set_anon_receive(flags[14], flags[15], true);
     new_acls.set_self_manage_list(flags[16]);
-    new_acls.set_self_manage_queue_limit(flags[17]);
+    new_acls.set_self_manage_send_queue_limit(flags[17]);
+    new_acls.set_self_manage_receive_queue_limit(flags[18]);
 
     if new_acls == acls {
         println!("{}", style("No changes made").yellow());