+++ affinidi-messaging-helpers/src/mediator_administration/main.rs
@@ -42,7 +42,8 @@ struct SharedConfig {
     pub mediator_did_hash: String,
     pub acl_mode: AccessListModeType,
     pub global_acl_default: MediatorACLSet,
-    pub queued_messages_soft: u32,
+    pub queued_send_messages_soft: i32,
+    pub queued_receive_messages_soft: i32,
 }
 
 impl SharedConfig {
@@ -105,17 +106,36 @@ impl SharedConfig {
             return Err("Couldn't find global_acl_default in Mediator Configuration".into());
         };
 
-        let queued_messages_soft = if let Some(queued_message_soft) = config
+        let queued_send_messages_soft = if let Some(queued_send_message_soft) = config
             .get("limits")
-            .and_then(|limits| limits.get("queued_messages_soft"))
+            .and_then(|limits| limits.get("queued_send_messages_soft"))
         {
-            if let Some(queued_message_soft) = queued_message_soft.as_u64() {
-                queued_message_soft as u32
+            if let Some(queued_send_message_soft) = queued_send_message_soft.as_i64() {
+                queued_send_message_soft as i32
             } else {
-                return Err("Couldn't find queued_messages_soft in Mediator Configuration".into());
+                return Err(
+                    "Couldn't find queued_send_messages_soft in Mediator Configuration".into(),
+                );
             }
         } else {
-            return Err("Couldn't find queued_messages_soft in Mediator Configuration".into());
+            return Err("Couldn't find queued_send_messages_soft in Mediator Configuration".into());
+        };
+
+        let queued_receive_messages_soft = if let Some(queued_receive_message_soft) = config
+            .get("limits")
+            .and_then(|limits| limits.get("queued_receive_messages_soft"))
+        {
+            if let Some(queued_receive_message_soft) = queued_receive_message_soft.as_i64() {
+                queued_receive_message_soft as i32
+            } else {
+                return Err(
+                    "Couldn't find queued_receive_messages_soft in Mediator Configuration".into(),
+                );
+            }
+        } else {
+            return Err(
+                "Couldn't find queued_receive_messages_soft in Mediator Configuration".into(),
+            );
         };
 
         Ok(SharedConfig {
@@ -124,7 +144,8 @@ impl SharedConfig {
             global_acl_default,
             our_admin_hash: String::new(),
             mediator_did_hash: digest(mediator_did),
-            queued_messages_soft,
+            queued_send_messages_soft,
+            queued_receive_messages_soft,
         })
     }
 }