+++ CHANGELOG.md
@@ -37,9 +37,9 @@ Missing versions on the changelog simply reflect minor deployment changes on our
 * JSON fields changed from UpperCamelCase to snake_case
   * Mediator Administration Protocol
   * Mediator Account Management
-* Queue limits can now be set per DID between a soft and hard limit
+* Queue limits can now be set per DID between a soft and hard limit, and separate for send/receive queues
   * Admin accounts can override and go above the hard limit as needed
-  * New ACL Flag enabled for can change queue_limit (SELF_MANAGE_QUEUE_LIMIT)
+  * New ACL Flag enabled for can change queue_limit (SELF_MANAGE_(SEND|RECEIVE)_QUEUE_LIMIT)
 * Ability to set an ephemeral header on messages that will not store the message
   * Instead, if the client is live-streaming it will send only via the live stream